SW algorithm upgrade, and fix UGENE-1407
git-svn-id: https://local.ugene.unipro.ru/svn/ugene/trunk@3633 7ca3a678-163c-e011-b84f-001e8c0267ff

+++ src/corelibs/U2Algorithm/src/smith_waterman/SmithWatermanSettings.h
@@ -48,8 +48,8 @@ static inline bool isDirect(StrandOption strand) {
 }
 
 struct GapModel {
-    float scoreGapOpen;
-    float scoreGapExtd;
+    int scoreGapOpen;
+    int scoreGapExtd;
 };
 
 struct SmithWatermanSettings {
